requirements.txt
asgiref==3.5.2
certifi==2022.5.18.1
charset-normalizer==2.0.12
Django==4.0.5
django-ckeditor==6.4.2
django-js-asset==2.0.0
django-social-share==2.2.1
idna==3.3
Pillow==9.1.1
requests==2.28.0
sqlparse==0.4.2
tinify==1.5.2
tzdata==2022.1
urllib3==1.26.9
